make_duration/requirements.txt
hsaudiotag3k==1.1.3
pandas==0.18.0
hsaudiotag==1.1.1
forked_path==0.2.3
